Fix autopan tap tempo

+++ Software/GuitarPedal/Util/audio_utilities.cpp
@@ -2,7 +2,7 @@
 
 float tempo_to_freq(uint32_t tempo)
 {
-    return tempo / 60.0f;
+  return static_cast<float>(tempo) / 60.0f;
 }
 
 uint32_t freq_to_tempo(float freq)
